pyproject.toml
[tool.poetry]
name = "marvel-e2e-pytest"
version = "0.1.0"
description = ""
authors = ["Victor Pires <vctprs@gmail.com>"]
readme = "README.md"
packages = [{include = "marvel_e2e_pytest"}]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "*"
requests = "*"
Faker = "*"
python-dotenv = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
